+++ NEWS.md
@@ -1,5 +1,5 @@
 # mascarade 0.3.1
-* fancyMask support setting colors
+* fancyMask support setting colors manually or to inherit from the main layer.
 
 # mascarade 0.3.0